chore: Cleanup NET8_0_OR_GREATER
