+++ src/Uno.UI.Toolkit/ElevatedView.Android.cs
@@ -20,11 +20,7 @@ partial class ElevatedView
 	private AndroidBitmap? _shadowBitmap;
 	private bool _invalidateShadow = true;
 
-#if NET8_0_OR_GREATER
 	protected override void DispatchDraw(Canvas canvas)
-#else
-	protected override void DispatchDraw(Canvas? canvas)
-#endif
 	{
 		if (Elevation > 0)
 		{
